Fix all clippy warnings to pass CI with -D warnings
- Auto-applied 26 clippy suggestions (redundant closures, useless vec, etc.)
- Added #![allow(dead_code)] to 10 modules with planned-but-not-yet-wired features
- Fixed &PathBuf -> &Path in tts.rs, .max().min() -> .clamp() in dedup.rs
- Merged identical if-blocks in knowledge_decay.rs and scoring/mod.rs
- Allowed proc-macro-generated manual_range_contains in scoring module
- Added targeted type_complexity and too_many_arguments allows

+++ src-tauri/src/digest_commands.rs
@@ -369,7 +369,7 @@ Rules:
                     &response.content,
                     Some(&llm_settings.model),
                     items.len(),
-                    Some(total_tokens as u64),
+                    Some(total_tokens),
                     Some(elapsed.as_millis() as u64),
                 ) {
                     error!(target: "4da::briefing", error = %e, "Failed to persist briefing to DB");
